water_calorie_bot/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
matplotlib==3.8.0
numpy==1.26.0